feat(a 60fps): Added a 60fps animation demo with particle effects and mouse interact...
# Project State

## Current Focus
Added a 60fps animation demo with particle effects and mouse interaction in the game loop example

## Context
This implements a visual showcase for the compositor's animation capabilities, demonstrating smooth rendering at 60fps with interactive particle effects triggered by mouse clicks

## Completed
- [x] Implemented particle system with physics (gravity, velocity)
- [x] Added mouse interaction for particle bursts
- [x] Created starfield background with twinkling effects
- [x] Implemented rocket animation with trail particles
- [x] Added turbo mode toggle (space key)
- [x] Included help display toggle (? key)
- [x] Set up proper terminal initialization/cleanup
- [x] Implemented FPS counter and timing control

## In Progress
- [ ] No active work in progress

## Blockers
- None identified

## Next Steps
1. Add sound effects for particle bursts
2. Implement score system based on click count
3. Add level progression or difficulty scaling
---
category: feat
scope: a 60fps
---

+++ src/compositor/engine.rs
@@ -289,7 +289,7 @@ impl Compositor {
                     let mut src_cell = plane.cells[src_idx];
 
                     if let Some(filter) = &plane.filter {
-                        filter.apply(&mut src_cell, abs_x, abs_y, render_time);
+                        filter.apply(&mut src_cell, abs_x, abs_y, render_time as f32);
                     }
 
                     blend_cells(&mut self.final_buffer[dest_idx], &src_cell, plane.opacity);
